Improve query logic
This should allow the plugin to report pending posts regardless of whether the Urgent flag has been defined.

The linkage to the Urgent flag is still brittle, relying on the actual value in our current website. This could be turned into an editable parameter, but that feels a bit like overengineering.